Fix autoboxing.
git-svn-id: http://svn.codehaus.org/xstream/trunk@803 9830eeb5-ddf4-0310-9ef7-f4b9a3e3227e

+++ xstream/src/java/com/thoughtworks/xstream/converters/extended/CGLIBEnhancedConverter.java
@@ -155,7 +155,7 @@ public class CGLIBEnhancedConverter implements Converter {
         reader.moveUp();
         if (reader.hasMoreChildren()) {
             reader.moveDown();
-            enhancer.setSerialVersionUID(Long.valueOf(reader.getValue()));
+            enhancer.setSerialVersionUID(Long.valueOf(reader.getValue()).longValue());
             reader.moveUp();
         }
         return enhancer.create();
